fixes #138, #134  issue with username and userslist view
username was not showing up because of a recent change, the fallback
condition to check for the username in case the name was missing from
the user's github.

In second issue, the check was missing to see if the users is coming
from github, we had fb & twitter users earlier. This also modifies the
query to filter out all other users other than github